Added cover photo ratios
Cover photo size is now based on one of three ratios: 4:3 (standard),
16:9 & 3:2. This can be set using the option coverratio.
Values used are calculated beforehand, for now.

The only difference in the "sender" field on the first page's Dutch
version and English version were the different words for P.O. Box and
The Netherlands. These are now set with the labels.
This allows the "drawing code"  part to be smaller.